Inicio das rotas e add página de login
