+++ src/pages/components/AccessibilityPanel.tsx
