+++ src/pages/components/ui/resizable.tsx
