+++ src/pages/components/ui/slider.tsx
